--- generate_csv.py
@@ -1,36 +1,0 @@
-import os
-from google import genai
-from dotenv import load_dotenv
-
-load_dotenv()
-
-# Get API Key from environment variable
-API_KEY = os.getenv('GEMINI_API_KEY')
-
-# Initialize the Gemini client
-client = genai.Client(api_key=API_KEY)
-
-# Function to generate quotes using Gemini
-def generate_quotes(prompt):
-    quotes = []
-
-    # Request content from Gemini model
-    response = client.models.generate_content(
-        model="gemini-1.5-flash-latest",
-        contents=prompt  # Use the passed prompt for generating quotes
-    )
-
-    # Assuming response.text contains the quote generated by the model
-    quotes_list = response.text.strip().split("\n")  # Split into individual quotes
-
-    # Ensure there are at least 50 quotes (or as many as possible)
-    for quote in quotes_list[:50]:
-        quotes.append({"Quote": quote.strip()})  # Store each quote
-
-    return quotes
-
-# If you want to call it directly (for testing or other purposes)
-if __name__ == "__main__":
-    prompt = "Generate a motivational quote about success."
-    generated_quotes = generate_quotes(prompt)
-    print(generated_quotes)
